Input: Add Novatek NT36xxx touchscreen driver
This is a driver for the Novatek in-cell touch controller and
supports various chips from the NT36xxx family, currently
including NT36525, NT36672A, NT36676F, NT36772 and NT36870.

Functionality like wake gestures and firmware flashing is not
included: I am not aware of any of these DrIC+Touch combo
chips not including a non-volatile memory and it should be
highly unlikely to find one, since the touch firmware is
embedded into the DriverIC one, which is obviously necessary
to drive the display unit.

However, the necessary address for the firmware update
procedure was included into the address table in this driver
so, in the event that someone finds the need to implement it
for a reason or another, it will be pretty straightforward to.

This driver is lightly based on the downstream implementation [1].
[1] https://github.com/Rasenkai/caf-tsoft-Novatek-nt36xxx

Signed-off-by: AngeloGioacchino Del Regno <kholk11@gmail.com>

input: touchscreen: nt36xxx: temporarily disable PM functions

During touchscreen suspend, the following errors occur and after which the touchscreen wouldnt work on resume. So disable the PM functions temporarily.

[16363.867191] PM: suspend entry (s2idle)
[16363.871845] Filesystems sync: 0.004 seconds
[16363.872858] Freezing user space processes ... (elapsed 0.002 seconds) done.
[16363.875064] OOM killer disabled.
[16363.875070] Freezing remaining freezable tasks ... (elapsed 0.000 seconds) done.
[16363.875972] printk: Suspending console(s) (use no_console_suspend to debug)
[16363.971053] gpi a00000.dma-controller: Error in Transaction
[16363.971060] geni_i2c a98000.i2c: DMA txn failed:3
[16364.996080] geni_i2c a98000.i2c: I2C timeout gsi flags:0 addr:0x1
[16364.997575] nt36xxx_ts 14-0062: Cannot enter suspend!!
[16364.997593] PM: dpm_run_callback(): nt36xxx_i2c_suspend+0x0/0x70 [nt36xxx] returns -110
[16364.997598] PM: Device 14-0062 failed to suspend: error -110
[16364.997603] PM: Some devices failed to suspend, or early wake event detected
[16365.138394] OOM killer enabled.
[16365.138397] Restarting tasks ... done.
[16365.141644] PM: suspend exit

input: nt36xxx: fix pressure not being reported for 1 or 2 finger taps

Fixes issues in TS LIB based touchscreen applications such pmOS osk-sdl.